requirements.txt
Django >=3.2.14
django-bootstrap-v5 >= 1.0.11
tweepy >=4.10.1
requests >=2.28.1
bs4 == 0.0.1
lxml >= 4.6.4
sysnet-ruian == 1.0.0
djangorestframework >=3.14.0
openpyxl >= 3.0.10
stanza >= 1.4.2
torch >= 1.13.0
pyproj >= 3.4.0
shapely >= 1.8.5.post1
elasticsearch~=8.5.3
pynominatim >= 0.3.0
SQLAlchemy >= 1.4.44
django-apscheduler >= 0.6.2
gunicorn
python-dotenv>=0.21.0
sysnet-shopping==1.0.5

APScheduler~=3.9.1.post1
setuptools~=60.2.0